+++ src/assets/scripts/client/aircraft/AircraftModel.js
@@ -1305,8 +1305,8 @@ export default class AircraftModel {
         );
 
         const readback = {};
-        readback.log = `taxi to and hold short runway ${runwayModel.name}`;
-        readback.say = `taxi to and hold short runway ${radio_runway(runwayModel.name)}`;
+        readback.log = `taxi to and hold short of Runway ${runwayModel.name}`;
+        readback.say = `taxi to and hold short of Runway ${radio_runway(runwayModel.name)}`;
 
         return [true, readback];
     }
@@ -1494,12 +1494,6 @@ export default class AircraftModel {
                 if (elapsed > this.taxi_time) {
                     this.setFlightPhase(FLIGHT_PHASE.WAITING);
                     this.moveToRunway(runwayModel);
-                    speech_say([
-                        { type: 'text', content: AirportController.airport_get().radio.twr },
-                        { type: 'callsign', content: this },
-                        { type: 'text', content: `, holding short runway ${runwayModel.getRadioName()}` }
-                    ]);
-                    UiController.ui_log(`${AirportController.airport_get().radio.twr}, ${this.callsign}, holding short runway ${runwayModel.name}`);
                 }
 
                 break;